chore: refactor to reduce code duplication
