+++ app/components/UI/Bridge/useSourceTokens.ts
@@ -1,149 +1,12 @@
-import { useMemo } from 'react';
+import { TokenIWithFiatAmount, useTokensWithBalance } from './useTokensWithBalance';
 import { useSelector } from 'react-redux';
-import { Hex } from '@metamask/utils';
-import { TokenI } from '../Tokens/types';
-import { selectTokensBalances } from '../../../selectors/tokenBalancesController';
-import { selectSelectedInternalAccountAddress } from '../../../selectors/accountsController';
-import { selectNetworkConfigurations } from '../../../selectors/networkController';
-import { selectTokenMarketData } from '../../../selectors/tokenRatesController';
-import { selectCurrencyRates, selectCurrentCurrency } from '../../../selectors/currencyRateController';
-import { deriveBalanceFromAssetMarketDetails, sortAssets } from '../Tokens/util';
-import { selectTokenSortConfig } from '../../../selectors/preferencesController';
 import { selectSelectedSourceChainIds } from '../../../core/redux/slices/bridge';
-import { selectAccountTokensAcrossChains } from '../../../selectors/multichain';
+import { Hex } from '@metamask/utils';
 
-export type TokenIWithFiatAmount = TokenI & { tokenFiatAmount: number };
-
-interface CalculateFiatBalancesParams {
-  assets: TokenI[];
-  multiChainMarketData: ReturnType<typeof selectTokenMarketData>;
-  multiChainTokenBalance: ReturnType<typeof selectTokensBalances>;
-  networkConfigurationsByChainId: ReturnType<typeof selectNetworkConfigurations>;
-  multiChainCurrencyRates: ReturnType<typeof selectCurrencyRates>;
-  currentCurrency: string;
-  selectedAddress: Hex;
-}
-
-/**
- * Calculate the fiat value of tokens and their balances
- * Originally rom app/components/UI/Tokens/index.tsx
- * @param {Object} params - The parameters object
- * @param {TokenI[]} params.assets - Array of tokens to calculate balances for
- * @param {Object} params.multiChainMarketData - Market data for tokens across chains
- * @param {Object} params.multiChainTokenBalance - Token balances across chains
- * @param {Object} params.networkConfigurationsByChainId - Network configurations by chain ID
- * @param {Object} params.multiChainCurrencyRates - Currency conversion rates across chains
- * @param {string} params.currentCurrency - Current currency code
- * @param {Hex} params.selectedAddress - Selected account address
- * @returns {Array<{tokenFiatAmount: number, balance: string, balanceFiat: string}>} Array of token balances with fiat values
- * @example
- * // Returns array of objects with:
- * // tokenFiatAmount: A sortable fiat value in the user's currency, e.g. 100.12345
- * // balance: A truncated non-atomic balance, e.g. 1.23456
- * // balanceFiat: A formatted fiat value, e.g. "$100.12345", "100.12345 cad"
- */
-export const calculateBalances = ({
-  assets,
-  multiChainMarketData,
-  multiChainTokenBalance,
-  networkConfigurationsByChainId,
-  multiChainCurrencyRates,
-  currentCurrency,
-  selectedAddress
-}: CalculateFiatBalancesParams): {
-  tokenFiatAmount: number;
-  balance: string;
-  balanceFiat: string;
-}[] =>
-  assets.map((token) => {
-    const chainId = token.chainId as Hex;
-    const multiChainExchangeRates = multiChainMarketData?.[chainId];
-    const multiChainTokenBalances =
-      multiChainTokenBalance?.[selectedAddress]?.[
-        chainId
-      ];
-    const nativeCurrency =
-      networkConfigurationsByChainId[chainId].nativeCurrency;
-    const multiChainConversionRate =
-      multiChainCurrencyRates?.[nativeCurrency]?.conversionRate || 0;
-
-    if (token.isETH || token.isNative) {
-      return {
-        tokenFiatAmount: parseFloat(token.balance) * multiChainConversionRate,
-        balance: token.balance,
-        balanceFiat: token.balanceFiat,
-      };
-    }
-
-    const res = deriveBalanceFromAssetMarketDetails(
-      token,
-      multiChainExchangeRates || {},
-      multiChainTokenBalances || {},
-      multiChainConversionRate || 0,
-      currentCurrency || '',
-    );
-    return {
-      tokenFiatAmount: res.balanceFiatCalculation ?? 0, // A sortable fiat value in the user's currency, e.g. 100.12345
-      balance: res.balance ?? '0', // A truncated non-atomic balance, e.g. 1.23456
-      balanceFiat: res.balanceFiat, // A formatted fiat value, e.g. "$100.12345", "100.12345 cad"
-    };
-  });
-
-// TODO Look into useMultichainBalances hook, or useGetFormattedTokensPerChain hook
+// TODO get top tokens from BridgeController
 export const useSourceTokens: () => TokenIWithFiatAmount[] = () => {
-  const tokenSortConfig = useSelector(selectTokenSortConfig);
-  const currentCurrency = useSelector(selectCurrentCurrency);
-  const selectedInternalAccountAddress = useSelector(selectSelectedInternalAccountAddress) as Hex;
-  const networkConfigurationsByChainId = useSelector(
-    selectNetworkConfigurations,
-  );
-
-  // Fiat conversion rates
-  const multiChainMarketData = useSelector(selectTokenMarketData);
-  const multiChainCurrencyRates = useSelector(selectCurrencyRates);
-
-  // All tokens across chains and their balances
-  // Includes native and non-native tokens
-  const accountTokensAcrossChains = useSelector(selectAccountTokensAcrossChains);
-  const multiChainTokenBalance = useSelector(selectTokensBalances);
-
-  // Chain ids to filter by
   const selectedSourceChainIds = useSelector(selectSelectedSourceChainIds);
+  const tokensWithBalance = useTokensWithBalance({ chainIds: selectedSourceChainIds as Hex[] });
 
-  const sortedTokens = useMemo(() => {
-    const allAccountTokens = (
-      Object.values(accountTokensAcrossChains).flat() as TokenI[]
-    )
-    .filter((token) => selectedSourceChainIds.includes(token.chainId as Hex));
-
-    const balances = calculateBalances({
-      assets: allAccountTokens,
-      multiChainMarketData,
-      multiChainTokenBalance,
-      networkConfigurationsByChainId,
-      multiChainCurrencyRates,
-      currentCurrency,
-      selectedAddress: selectedInternalAccountAddress,
-    });
-    const properTokens = allAccountTokens.map((token, i) => ({
-      ...token,
-      tokenFiatAmount: balances[i].tokenFiatAmount ?? 0,
-      balance: balances[i].balance,
-      balanceFiat: balances[i].balanceFiat,
-      symbol: token.isETH ? 'ETH' : token.symbol, // TODO: not sure why symbol is ETHEREUM, will also break the token icon for ETH
-    }));
-    return sortAssets(properTokens, tokenSortConfig);
-  }, [
-    accountTokensAcrossChains,
-    multiChainMarketData,
-    multiChainTokenBalance,
-    networkConfigurationsByChainId,
-    multiChainCurrencyRates,
-    currentCurrency,
-    selectedInternalAccountAddress,
-    tokenSortConfig,
-    selectedSourceChainIds,
-  ]);
-
-  return sortedTokens;
+  return tokensWithBalance;
 };
